scaleatt/requirements.txt
opencv_python==4.7.0.68
numpy==1.24.1
tensorflow==2.11.0
scikit_image==0.19.3
matplotlib==3.6.3
cvxpy==1.3.0
Keras==2.11.0
Pillow==9.4.0
scikit_image==0.15.0
scikit_learn==1.2.1
scipy==1.10.0
pandas==1.5.3
fpdf==1.7.2
cython==0.29.33
seaborn==0.10.0
argparse==1.4.0